Use size_t consistently for memory region size
Ignore-this: 30fadcb1c5d2ca9f5ced0ad266c60da4

darcs-hash:e5babf33003f9ab5ebef8e7789458e29988c7faf

+++ cbits/HsMmap.h
@@ -1,11 +1,13 @@
 #ifndef __HSMMAP_H__
 #define __HSMMAP_H__
 
+#include <stddef.h>
+
 void *system_io_mmap_file_open(const char *filepath, int mode);
 
 void system_io_mmap_file_close(void *handle);
 
-void *system_io_mmap_mmap(void *handle, int mode, long long offset, int size);
+void *system_io_mmap_mmap(void *handle, int mode, long long offset, size_t size);
 
 void system_io_mmap_munmap(void *sizeasptr, void *ptr);
